assign correct ownerKeyVersion when updating calendar event
Before making the REST request to update a calendar event we assign some technical attributes like _id and _ownerEncSessionKey to the new/updated event. The assignment of the _ownerKeyVersion was missing wich is required if we have another version of the group key. It is suspicious that we have to do this assignement at all because before editing the event we clone the existing entity and all attributes should have been assigned to the new one already. It should probably also not the responsibility of the updateCalendarEvent function to assign these attributes. This commit just fixes this issue itself so the user does not get the error but we should review the clone and update mechanism.

fixes #7533